added a subtraction feature and made the total for division a float

+++ src/main/java/ie/atu/Calculator.java
@@ -38,10 +38,10 @@ public class Calculator {
     {
         System.out.println("Please enter your first number: ");
         Scanner inputs = new Scanner(System.in);
-        int firstNumber = inputs.nextInt();
+        float firstNumber = inputs.nextInt();
 
         System.out.println("Please enter your second number: ");
-        int secondNumber = inputs.nextInt();
+        float secondNumber = inputs.nextInt();
 
         float total = firstNumber / secondNumber;
         System.out.println("The total is: " + total);
